+++ lib/src/utils/flutter_types_utils.dart
@@ -15,6 +15,9 @@ bool isWidgetOrSubclass(DartType? type) =>
 bool isRenderObjectOrSubclass(DartType? type) =>
     _isRenderObject(type) || _isSubclassOfRenderObject(type);
 
+bool isRenderObjectWidgetOrSubclass(DartType? type) =>
+    _isRenderObjectWidget(type) || _isSubclassOfRenderObjectWidget(type);
+
 bool isWidgetStateOrSubclass(DartType? type) =>
     _isWidgetState(type) || _isSubclassOfWidgetState(type);
 
@@ -70,3 +73,9 @@ bool _isRenderObject(DartType? type) =>
 
 bool _isSubclassOfRenderObject(DartType? type) =>
     type is InterfaceType && type.allSupertypes.any(_isRenderObject);
+
+bool _isRenderObjectWidget(DartType? type) =>
+    type?.getDisplayString(withNullability: false) == 'RenderObjectWidget';
+
+bool _isSubclassOfRenderObjectWidget(DartType? type) =>
+    type is InterfaceType && type.allSupertypes.any(_isRenderObjectWidget);